api/requirements.txt
python-dotenv
openai<1.0.0
pinecone-client
transformers
sentence_transformers
gunicorn==20.1.0
flask
flask-cors
flask-socketio
